Restaurant

+++ Scripts_C#/Resto.MVC/Controllers/RestauranteController.cs
@@ -63,7 +63,7 @@ public class RestauranteController : Controller
         return RedirectToAction(nameof(ObtenerRestaurants));
     }
 [HttpGet]
-public async Task<IActionResult> ObtenerDetalleResto(int idRestaurant)
+public async Task<IActionResult> ObtenerDetalleResto(ushort idRestaurant)
 {
     var resto = await Ado.DetalleRestaurantAsync(idRestaurant);
